Bit of refactoring to tidy up
